refactor: fix app.test.js
