+++ backend/__tests__/app.test.js
@@ -12,7 +12,7 @@ describe("app", () => {
   });
 
   test("app responds with status 500 if an unexpected error occurs", async () => {
-    vi.doMock("../utils/rateLimiter.js", () => ({
+    vi.doMock("../src/utils/rateLimiter.js", () => ({
       global: vi.fn((req, res, next) => next()),
       api: vi.fn(() => {
         throw new Error("Unexpected error");
