+++ backend/src/app.ts
@@ -73,6 +73,7 @@ app.use((err: Error, _req: Request, res: Response, _next: NextFunction) => {
   logger.error(err);
 
   return sendError(res, {
+    status: 500,
     message: "Server error: please try again later.",
   });
 });
